refactor: . Move DateAbbreviations.ts and its tests to DateTime/ sub-directory
